+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Power Types
 - Added `action_when_lightning_struck` power type.
+- Added `damage_nearby_on_hit` and `damage_nearby_when_hit` power types.
 
 ### Config
 - Removed `performVersionCheck` config setting and its related code from the mod.